+++ lua/obsidian/async.lua
@@ -3,7 +3,7 @@ local async = require "plenary.async"
 local channel = require("plenary.async.control").channel
 local log = require "obsidian.log"
 local util = require "obsidian.util"
-local iter, uv = vim.iter, vim.uv
+local uv = vim.uv
 
 local M = {}
 
@@ -246,83 +246,6 @@ ThreadPoolExecutor.submit = function(self, fn, callback, ...)
     end
   end)
   ctx:queue(...)
-end
-
----Represents a file.
----@class obsidian.File : obsidian.ABC
----@field fd userdata
-local File = abc.new_class()
-
-M.File = File
-
----@param path string|obsidian.Path
----@param mode string|?
----@return obsidian.File
-File.open = function(path, mode)
-  local self = File.init()
-  local err, fd = async.uv.fs_open(tostring(path), mode and mode or "r", 438)
-  assert(not err, err)
-  self.fd = fd
-  return self
-end
-
----Close the file.
----@param self obsidian.File
-File.close = function(self)
-  local err = async.uv.fs_close(self.fd)
-  assert(not err, err)
-end
-
----Get at iterator over lines in the file.
----@param include_new_line_char boolean|?
-File.lines = function(self, include_new_line_char)
-  local offset = 0
-  local chunk_size = 1024
-  local buffer = ""
-  local eof_reached = false
-
-  local lines = function()
-    local idx_s, idx_e = string.find(buffer, "\r?\n")
-    while idx_s == nil and not eof_reached do
-      ---@diagnostic disable-next-line: redefined-local
-      local err, data
-      err, data = async.uv.fs_read(self.fd, chunk_size, offset)
-      assert(not err, err)
-      if string.len(data) == 0 then
-        eof_reached = true
-      else
-        buffer = buffer .. data
-        offset = offset + string.len(data)
-        idx_s, idx_e = string.find(buffer, "\r?\n")
-      end
-    end
-
-    if idx_s ~= nil then
-      assert(idx_e)
-      local line = string.sub(buffer, 1, idx_s)
-      buffer = string.sub(buffer, idx_e + 1)
-      if include_new_line_char then
-        return line
-      else
-        return string.sub(line, 1, -2)
-      end
-    else
-      return nil
-    end
-  end
-
-  return lines
-end
-
----Write all lines to the file.
----@param lines string[]|function
-File.write_lines = function(self, lines)
-  for line in iter(lines) do
-    if not string.find(line, "[\r\n]") then
-      line = line .. "\n"
-    end
-    async.uv.fs_write(self.fd, line)
-  end
 end
 
 ---@param cmds string[]